lab4/requirements.txt
requests>=2.31.0,<3
telethon>=1.34.0,<2
python-dotenv>=1.0.1,<2
matplotlib>=3.8.0,<4
pandas>=2.2.0,<3

